+++ playground/vite/downwind.config.ts
@@ -3,6 +3,6 @@ import type { DownwindConfig } from "../../src/types";
 export const config: DownwindConfig = {
   shortcuts: {
     "btn": "py-2 px-4 font-semibold rounded-lg shadow-md",
-    "btn-green": "text-white bg-green-500 hover:bg-green-700",
+    "btn-green": "text-white bg-green-500 hover:bg-green-600",
   },
 };