+++ playground/vite/src/App.tsx
@@ -8,12 +8,12 @@ export const App = () => {
   return (
     <div className={styles.container}>
       <h1 className="font-bold text-xl">Vite + React</h1>
-      <div className="p-4">
+      <div className="p-4 text-lg">
         <button
           className="btn btn-green"
           onClick={() => setCount((v) => v + 1)}
         >
-          count is {count}
+          count is <span className="tabular-nums">{count}</span>
         </button>
         <p className="mt-4">
           Edit <code>src/App.tsx</code> and save to test HMR