+++ tests/codegen.test.ts
@@ -1,7 +1,9 @@
 import { snapshotTest } from "./test-utils";
 
-snapshotTest("codegen", (downwind) => downwind.codegen({ omitContent: false }));
+snapshotTest("codegen", (downwind) =>
+  downwind.codegen({ mode: "WITH_CONTENT" }),
+);
 
 snapshotTest("codegen-omit", (downwind) =>
-  downwind.codegen({ omitContent: true }),
+  downwind.codegen({ mode: "OMIT_CONTENT" }),
 );